+++ parley-service-jpa/src/main/java/org/sidequest/parley/repository/ChatRoomUsersRepository.java
@@ -1,5 +1,6 @@
 package org.sidequest.parley.repository;
 
+import org.sidequest.parley.entity.ChatRoomUsersEntity;
 import org.sidequest.parley.model.ChatRoomUsers;
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -7,5 +8,5 @@ import java.util.Optional;
 
 
 public interface ChatRoomUsersRepository extends JpaRepository<ChatRoomUsersEntity, Long> {
-    Optional<ChatRoomUsers> findByUserIdAndChatRoomId(int userId, int chatRoomId);
+    Optional<ChatRoomUsers> findByUserIdAndChatRoomId(Long userId, Long chatRoomId);
 }
